Change date folder structure for home-page. Move ITProfession file to data/home-page. Remove unnessesery data from this file.
